fix: corrigido erros nas configurações de cliente que impediam o sistema de subir

+++ vSpring/autoagenda/src/main/java/br/com/autoagenda/autoagenda/controller/C_Cliente.java
@@ -34,7 +34,7 @@ public class C_Cliente {
     }
 
 	@PostMapping("/apagarCliente")
-    public String apagarCliente(@RequestParam Long idCliente) {
+    public String apagarCliente(@RequestParam Integer idCliente) {
 		if(idCliente != null) { repo.deleteById(idCliente); }
 		return "mensagem de retorno de deu bom";
     }
